% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defer}
\alias{defer}
\title{Defer doing anything to data}
\usage{
defer(...)
}
\arguments{
\item{...}{Named arguments}
}
\value{
A \code{data.frame}
}
\description{
This function is used as a stand-in for an element modifying function that
does nothing special to the data except place it in a data.frame.
}
\examples{
element_line_seq(fun = defer)
}
