% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_line_seq.R
\name{element_line_seq}
\alias{element_line_seq}
\alias{element_grob.element_line_seq}
\title{Interpolated line theme element}
\usage{
element_line_seq(
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE,
  fun = defer,
  n = 50
)
}
\arguments{
\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{linewidth}{Line/border size in mm.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{fun}{A \code{function} to apply to the interpolated lines}

\item{n}{An \code{integer} of length one indicating how many points to
interpolate.}

\item{x}{A single number specifying size relative to parent element.}
}
\value{
A \code{list} of the class \code{element_line_seq}
}
\description{
Interpolates a line.
}
\examples{
element_line_seq()
}
\keyword{internal}
