% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_polygon.R
\name{element_polygon}
\alias{element_polygon}
\title{Polygon theme elements}
\usage{
element_polygon(
  fill = NULL,
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  linejoin = NULL,
  lineend = NULL,
  color = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{fill}{A colour specification for the fill.}

\item{colour}{A colour specification for the borders or lines.}

\item{linewidth}{A \code{numeric} for the width of lines in millimetres.}

\item{linetype}{A \code{integer} or \code{string} specifying a line type.}

\item{linejoin}{One of the following strings: \code{"round"}, \code{"mitre"}
or \code{"bevel"}.}

\item{lineend}{One of the following strings: \code{"round"}, \code{"mitre"}
or \code{"bevel"}.}

\item{color}{Alias for the \code{colour} argument.}

\item{inherit.blank}{A \code{logical} of length 1: should this element
inherit the existence of an \code{element_blank} among its parents? If
\code{TRUE}, the existance of a blank element among its parents will cause
this element to be blank as well. If \code{FALSE}, any blank parent will be
ignored when calculating final element state.}
}
\value{
A \code{list} with the class \code{element_polygon}.
}
\description{
Constructor for polygon theme elements.
}
\details{
Themes in vanilla ggplot2 do not contain polygons.
}
\examples{
NULL
}
