% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_rect_round.R
\name{element_rect_round}
\alias{element_rect_round}
\alias{element_grob.element_rect_round}
\title{Rounded rectangle theme element}
\usage{
element_rect_round(
  fill = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE,
  radius = unit(0.1, "snpc")
)
}
\arguments{
\item{fill}{Fill colour.}

\item{colour}{Line/border colour. Color is an alias for colour.}

\item{size}{Line/border size in mm; text size in pts.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{color}{Line/border colour. Color is an alias for colour.}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{radius}{A \code{unit} of length 1 specifying the maximum radius of
circles inscribing the corners. If given a \code{numeric}, it is converted
to a \code{unit} object with the \code{"snpc"} unit.}

\item{x}{A single number specifying size relative to parent element.}

\item{lineend}{Line end Line end style (round, butt, square)}
}
\value{
A \code{list} of the class \code{element_rect_round}.
}
\description{
Draws rectangles with rounded corners in conjuction with the theme system.
}
\examples{
# Adding rounded rectangles to geoms and theme elements
ggplot(mpg, aes(class, colour = class)) +
  geom_bar_theme(element = element_rect_round(radius = 0.25)) +
  facet_grid(~ year) +
  theme(
    strip.background = element_rect_round(radius = 0.5),
    panel.background = element_rect_round()
  )
}
\keyword{internal}
