% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_rect_seq.R
\name{element_rect_seq}
\alias{element_rect_seq}
\alias{element_grob.element_rect_seq}
\title{Interpolated rectangle theme element.}
\usage{
element_rect_seq(
  fill = NULL,
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE,
  fun = defer,
  n = 50,
  sides = "tblr"
)
}
\arguments{
\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{linewidth}{Line/border size in mm.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}

\item{fun}{A \code{function} to apply to the interpolated lines}

\item{n}{An \code{integer} of length one indicating how many points to
interpolate.}

\item{sides}{A \code{character} of length one containing any of \code{"t"},
\code{"l"}, \code{"b"} and/or \code{"r"}. If these letters are present,
edges are drawn at the top (t), left (l), bottom (b) or right (r),
respectively. Including all these letters will draw all rectangle edges
(default), whereas including none of these letters will not draw edges.}

\item{x}{A single number specifying size relative to parent element.}

\item{lineend}{Line end Line end style (round, butt, square)}
}
\value{
A \code{list} of the class \code{element_line_seq}
}
\description{
Interpolates the sides of a rectangle.
}
\examples{
element_rect_seq()
}
\keyword{internal}
