% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vctrs.R
\name{utils_vctrs}
\alias{utils_vctrs}
\title{Vctr utilies}
\usage{
\method{vec_arith}{defaulted}(op, x, y, ...)

\method{vec_arith}{grouped_variable}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Elementalist uses vctrs to keep track of some circumstances in
  which variables were created, namely when they were defaulted and wheter
  some variables came from the theme or the geom. These functions help
  organise this a bit better, but should not be of interest to any user. This
  documentation merely exists to satisfy the R CMD check.
}
\keyword{internal}
