% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_wiggle.R
\name{wiggle}
\alias{wiggle}
\alias{element_line_wiggle}
\alias{element_rect_wiggle}
\title{Wiggle style}
\usage{
wiggle(amount = 3, seed = NULL)

element_line_wiggle(amount = 3, seed = NULL, ...)

element_rect_wiggle(amount = 3, seed = NULL, ...)
}
\arguments{
\item{amount}{A \code{numeric} of length 1 setting the amount of wiggling to
occur.}

\item{seed}{An \code{integer} to set the seed for reproducible wiggling.}

\item{...}{
  Arguments passed on to \code{\link[=element_line_seq]{element_line_seq}}, \code{\link[=element_rect_seq]{element_rect_seq}}
  \describe{
    \item{\code{fun}}{A \code{function} to apply to the interpolated lines}
    \item{\code{n}}{An \code{integer} of length one indicating how many points to
interpolate.}
    \item{\code{colour}}{Line/border colour. Color is an alias for colour.}
    \item{\code{size}}{Line/border size in mm; text size in pts.}
    \item{\code{linetype}}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}
    \item{\code{lineend}}{Line end Line end style (round, butt, square)}
    \item{\code{color}}{Line/border colour. Color is an alias for colour.}
    \item{\code{arrow}}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}}
    \item{\code{inherit.blank}}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
    \item{\code{sides}}{A \code{character} of length one containing any of \code{"t"},
\code{"l"}, \code{"b"} and/or \code{"r"}. If these letters are present,
edges are drawn at the top (t), left (l), bottom (b) or right (r),
respectively. Including all these letters will draw all rectangle edges
(default), whereas including none of these letters will not draw edges.}
  }}
}
\value{
A \code{function}.
}
\description{
'Wiggle' is a theme style that adds an amount of cumulative uniform noise to
interpolated lines, making them wiggle a bit.
}
\details{
The amount of wiggle added to lines and rectangles is absolute. This
makes it easier to make more uniform wiggles, but causes relative distortion
when resizing the plot window or device.

The \code{element_*_wiggle()} functions are convenience wrappers around
\code{element_*_seq()}.

The \code{wiggle()} function is a function factory that produces a
function, that subsequently can be used to wiggle points.
}
\examples{
plot <- ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  facet_wrap(~ drv)

# Line wiggles
plot + theme(
  axis.line = element_line_wiggle(1),
  panel.grid.major = element_line_wiggle(1)
)

# Turning clipping off makes the panel background wiggle better
plot +
  coord_cartesian(clip = "off") +
  theme(
    strip.background = element_rect_wiggle(),
    panel.background = element_rect_wiggle(sides = "lb", colour = "blue")
  )
}
